+++ src/deprecated.jl
@@ -1,0 +1,1 @@
+@deprecate _array_value(x) internal_value(x);