+++ test/deprecated.jl
@@ -1,0 +1,7 @@
+using ManifoldsBase
+
+@testset "deprecated functions still working" begin
+    x = [1.0, 2.0, 3.0]
+    y = ValidationMPoint(x)
+    @test ManifoldsBase._array_value(y) == internal_value(y)
+end